pyproject.toml
[tool.poetry]
name = "elastic"
version = "0.1.0"
description = ""
authors = ["Kiran Kumar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
elasticsearch = "^8.14.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
